move the cli parser definitions
