+++ src/plotman/_tests/job_test.py
@@ -1,14 +1,8 @@
-import contextlib
-import datetime
-import locale
 import importlib.resources
-import os
 import pathlib
-import typing
 
 import pendulum
 import pytest
-from plotman import job
 from plotman._tests import resources
