+++ src/plotman/job.py
@@ -1,25 +1,15 @@
-# TODO do we use all these?
-import argparse
 import contextlib
 import functools
-import logging
 import os
-import random
-import re
-import sys
 import glob
 import time
 from datetime import datetime
-from enum import Enum, auto
-from subprocess import call
 import typing
 
 import attr
 import click
-import pendulum
 import psutil
 
-from plotman import chia, madmax
 import plotman.errors
 if typing.TYPE_CHECKING:
     import plotman.errors
